Seperated some of the functions
