+++ external music functions/deleteNowPlaying.js
@@ -1,0 +1,7 @@
+module.exports = async function deleteNowPlaying(server) {
+    try {
+        await server.nowPlaying.delete();
+    } catch (err) {
+        console.log(err);
+    }
+}
