+++ external music functions/serverMap.js
@@ -1,0 +1,5 @@
+const serverMap = new Map();
+
+module.exports = {
+    serverMap: serverMap
+}
