+++ help functions/musicHelp.js
@@ -19,7 +19,7 @@ module.exports = function musicHelp(interaction) {
             { name: 'Skipping Music', value: 'Say you\'re listening to song a song that you\'re just not digging anymore. Well all you have to do is say "/skip" and I\'ll go to the next song in the queue! But just know this command won\'t work if there\'s nothing for me to skip to.', inline: true},
             { name: 'Previous track', value: 'Did you skip a track that you were actually digging? Well that was dumb. But luckily I have a command for that too. Just say "/prev" and I\'ll got back to the previous track. But just like the skip command, this command won\'t work if there\'s nothing for me to go back to.', inline: true},
             { name: '\u200B', value: '\u200B' },
-            { name: 'See the queue', value: 'Did you already forget what songs you put into the queue? Well I guess if you want I can show you what\'s coming up next. Just say "/queue" and I will show you the last song I played, the song that I am currently playing, and since you asked nicely I will show you the next 10 songs. Sound good? 👍', inline: true },
+            { name: 'See the queue', value: 'Did you already forget what songs you put into the queue? Well I guess if you want I can show you what\'s coming up next. Just say "/queue" and I will show you the last song I played, the song that I am currently playing, and since you asked nicely I will show you the next 10 songs. Sound good? 👍 I\'ll also show you if the queue is set to loop or not 😁', inline: true },
             { name: 'Loop the queue', value: 'Do you hate it when the music stops? yeah so do I, but if you don\'t want the music to stop you can have the queue on an infinite loop! Just say "/loop" and the music will just keep playing... Forever... And ever... Or something. But just know that this will make the /previous function not work anymore 🤷‍♀️ ', inline: true },
             { name: '\u200B', value: '\u200B' }
         )
